feat(admin): create learning content test method generic
